fix comment

+++ oneflow/user/kernels/dim_gather_kernel_util.cu
@@ -24,7 +24,7 @@ namespace user_op {
 template<typename IN_T, typename IDX_T>
 __global__ void DoCUDADimGather(const DimOpIndexNdHelper<IDX_T> input_nd_helper,
                                 const DimOpIndexNdHelper<IDX_T> index_nd_helper, int ndim,
-                                int64_t elem_cnt, int64_t dim_length, int32_t dim, const IDX_T* index,
+                                int64_t elem_cnt, int32_t dim_length, int32_t dim, const IDX_T* index,
                                 const IN_T* input, IN_T* output) {
   DoDimGather<IN_T, IDX_T>(input_nd_helper, index_nd_helper, ndim, elem_cnt, dim_length, dim, index, input,
                            output);
@@ -33,7 +33,7 @@ __global__ void DoCUDADimGather(const DimOpIndexNdHelper<IDX_T> input_nd_helper,
 template<typename IDX_T, typename IN_T>
 struct DimGatherFunctor<DeviceType::kCUDA, IN_T, IDX_T> final {
   void operator()(ep::Stream* stream, const DimOpIndexNdHelper<IDX_T>& input_nd_helper,
-                  const DimOpIndexNdHelper<IDX_T>& index_nd_helper, int ndim, int64_t elem_cnt, int64_t dim_length, 
+                  const DimOpIndexNdHelper<IDX_T>& index_nd_helper, int ndim, int64_t elem_cnt, int32_t dim_length, 
                   int32_t dim, const IDX_T* index, const IN_T* input, IN_T* output) {
     RUN_CUDA_KERNEL((DoCUDADimGather<IN_T, IDX_T>), stream, BlocksNum4ThreadsNum(elem_cnt),
                     input_nd_helper, index_nd_helper, ndim, elem_cnt, dim_length, dim, index, input, output);
@@ -44,7 +44,7 @@ struct DimGatherFunctor<DeviceType::kCUDA, IN_T, IDX_T> final {
 template<typename IDX_T>
 struct DimGatherFunctor<DeviceType::kCUDA, float16, IDX_T> final {
   void operator()(ep::Stream* stream, const DimOpIndexNdHelper<IDX_T>& input_nd_helper,
-                  const DimOpIndexNdHelper<IDX_T>& index_nd_helper, int ndim, int64_t elem_cnt, int64_t dim_length,
+                  const DimOpIndexNdHelper<IDX_T>& index_nd_helper, int ndim, int64_t elem_cnt, int32_t dim_length,
                   int32_t dim, const IDX_T* index, const float16* input, float16* output) {
     RUN_CUDA_KERNEL((DoCUDADimGather<half, IDX_T>), stream, BlocksNum4ThreadsNum(elem_cnt),
                     input_nd_helper, index_nd_helper, ndim, elem_cnt, dim_length, dim, index,
