+++ oneflow/user/kernels/dim_gather_kernel_util.h
@@ -43,14 +43,14 @@ template<DeviceType device_type, typename IN_T, typename IDX_T>
 struct DimGatherFunctor final {
   void operator()(ep::Stream* stream, const DimOpIndexNdHelper<IDX_T>& input_nd_helper,
                   const DimOpIndexNdHelper<IDX_T>& index_nd_helper, int ndim, int64_t elem_cnt,
-                  int64_t dim_length, int32_t dim, const IDX_T* index, const IN_T* input,
+                  int32_t dim_length, int32_t dim, const IDX_T* index, const IN_T* input,
                   IN_T* output);
 };
 
 template<typename IN_T, typename IDX_T>
 OF_DEVICE_FUNC void DoDimGather(const DimOpIndexNdHelper<IDX_T>& input_nd_helper,
                                 const DimOpIndexNdHelper<IDX_T>& index_nd_helper, int ndim,
-                                int64_t elem_cnt, int64_t dim_length, int32_t dim,
+                                int64_t elem_cnt, int32_t dim_length, int32_t dim,
                                 const IDX_T* index, const IN_T* input, IN_T* output) {
   XPU_1D_KERNEL_LOOP(index_offset, elem_cnt) {
     IDX_T coordinate[kDimGatherMaxDimCount] = {0};
